+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ## Unreleased
 
 * Remove option select GA4 attributes ([PR #3625](https://github.com/alphagov/govuk_publishing_components/pull/3625))
+* Fix various bugs with the GA4 pageview tracker ([PR #3626](https://github.com/alphagov/govuk_publishing_components/pull/3626))
 
 ## 35.16.0